refactor(packet): move `IpProtocol` to `packet.rs`
