+++ src/tracing/packet/ipv4.rs
@@ -1,5 +1,5 @@
 use crate::tracing::packet::buffer::Buffer;
-use crate::tracing::packet::fmt_payload;
+use crate::tracing::packet::{fmt_payload, IpProtocol};
 use std::fmt::{Debug, Formatter};
 use std::net::Ipv4Addr;
 
@@ -16,34 +16,6 @@ const PROTOCOL_OFFSET: usize = 9;
 const CHECKSUM_OFFSET: usize = 10;
 const SOURCE_OFFSET: usize = 12;
 const DESTINATION_OFFSET: usize = 16;
-
-///
-#[derive(Debug, Clone, Copy, Eq, PartialEq)]
-pub enum IpProtocol {
-    Icmp,
-    IcmpV6,
-    Udp,
-    Tcp,
-    Other(u8),
-}
-
-impl IpProtocol {
-    #[must_use]
-    pub fn id(self) -> u8 {
-        match self {
-            Self::Icmp => 1,
-            Self::IcmpV6 => 58,
-            Self::Udp => 17,
-            Self::Tcp => 6,
-            Self::Other(id) => id,
-        }
-    }
-
-    #[must_use]
-    pub fn new(value: u8) -> Self {
-        Self::Other(value)
-    }
-}
 
 /// Represents an IPv4 Packet.
 ///
